+++ bin/installed_libraries.ml
@@ -13,12 +13,11 @@ let term =
       & info [ "na"; "not-available" ]
           ~doc:"List libraries that are not available and explain why")
   in
-  Common.set_common common;
+  let config = Common.set_common common in
   let capture_outputs = Common.capture_outputs common in
-  Scheduler.go ~common (fun () ->
+  Scheduler.go ~common ~config (fun () ->
       let open Fiber.O in
       let* (_env : Env.t) = Import.Main.setup_env ~capture_outputs in
-      let* () = Memo.Build.run (Workspace.init ()) in
       let* ctxs = Memo.Build.run (Context.DB.all ()) in
       let ctx = List.hd ctxs in
       let findlib = ctx.findlib in