+++ bin/ocaml_merlin.ml
@@ -29,18 +29,10 @@ let term =
              ouptut.")
   in
   let common = Common.set_print_directory common false in
-  Common.set_common common ~log_file:No_log_file;
-  Scheduler.go ~common (fun () ->
+  let config = Common.set_common common ~log_file:No_log_file in
+  Scheduler.go ~common ~config (fun () ->
       Dune_engine.Source_tree.init ~recognize_jbuilder_projects:true
         ~ancestor_vcs:None;
-      let x = Common.x common in
-      let workspace_file =
-        Common.workspace_file common |> Option.map ~f:Arg.Path.path
-      in
-      let open Fiber.O in
-      let* () =
-        Memo.Build.run (Dune_rules.Workspace.init ?x ?workspace_file ())
-      in
       match dump_config with
       | Some s -> Dune_rules.Merlin_server.dump s
       | None -> Dune_rules.Merlin_server.start ())
@@ -74,18 +66,10 @@ module Dump_dot_merlin = struct
               "The path to the folder of which the configuration should be \
                printed. Defaults to the current directory.")
     in
-    Common.set_common common ~log_file:No_log_file;
-    Scheduler.go ~common (fun () ->
+    let config = Common.set_common common ~log_file:No_log_file in
+    Scheduler.go ~common ~config (fun () ->
         Dune_engine.Source_tree.init ~recognize_jbuilder_projects:true
           ~ancestor_vcs:None;
-        let x = Common.x common in
-        let workspace_file =
-          Common.workspace_file common |> Option.map ~f:Arg.Path.path
-        in
-        let open Fiber.O in
-        let* () =
-          Memo.Build.run (Dune_rules.Workspace.init ?x ?workspace_file ())
-        in
         match path with
         | Some s -> Dune_rules.Merlin_server.dump_dot_merlin s
         | None -> Dune_rules.Merlin_server.dump_dot_merlin ".")